% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_formulas.r
\name{extract_formulas}
\alias{extract_formulas}
\title{Helper function to extract and format formulas}
\usage{
extract_formulas(arguments, formulas)
}
\description{
Helper function to extract and format formulas
}
