% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faux_census.r
\docType{data}
\name{faux_census}
\alias{faux_census}
\title{A small sample of a fabricated census-like dataset}
\format{
A tibble with 20 rows and 6 variables.
}
\source{
Fabricated
}
\usage{
faux_census
}
\description{
A dataset containing fake demographic data, used in the \code{fauxnaif} vignette.
}
\examples{
faux_census
}
\keyword{datasets}
