% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if.r
\name{na_if}
\alias{na_if}
\alias{faux_na_if}
\alias{fauxnaif}
\title{Convert values to NA}
\usage{
na_if(x, ...)

faux_na_if(x, ...)
}
\arguments{
\item{x}{Vector to modify}

\item{...}{Values to replace with \code{NA}}
}
\value{
A modified version of \code{x} that replaces any values contained in \code{...}
with NA.
}
\description{
This is a replacement for \code{\link[dplyr:na_if]{dplyr::na_if()}}.
It is useful if you want to convert annoying values to \code{NA}.
Unlike \code{\link[dplyr:na_if]{dplyr::na_if()}}, this function allows you to specify multiple values
to be replaced with \code{NA} at the same time.
\code{faux_na_if()} is offered as an alternative to avoid clashes with \link{dplyr}.
}
\examples{
na_if(1:5, 2, 4)

y <- c("abc", "", "def", "NA", "ghi", 42, "jkl", "NULL", "mno")
na_if(y, "", c("NA", "NULL"), 1:100)

\dontrun{
# This function handles vector values differently than dplyr,
# and returns a different result with vector replacement values:
na_if(1:5, 5:1)
dplyr::na_if(1:5, 5:1)

# na_if is particularly useful inside mutate,
# and is meant for use with vectors rather than entire data frames
dplyr::starwars \%>\%
  select(name, hair_color) \%>\%
  mutate(hair_color = na_if(hair_color, "unknown", "none"))

# na_if can also be used with scoped variants of mutate
# like mutate_if to mutate multiple columns
dplyr::starwars \%>\%
  mutate_if(is.character, ~ na_if(., "unknown", "none"))
}
}
\seealso{
\code{\link[dplyr:na_if]{dplyr::na_if()}} to replace a single value with \code{NA}.

\code{\link[dplyr:coalesce]{dplyr::coalesce()}} to replace missing values with a specified value.

\code{\link[tidyr:replace_na]{tidyr::replace_na()}} to replace \code{NA} with a value.

\code{\link[dplyr:recode]{dplyr::recode()}} and \code{\link[dplyr:case_when]{dplyr::case_when()}} to more generally replace
values.
}
